+++ src/main/java/com/daemawiki/external/web/rest/auth/AuthDTOMapper.java
@@ -1,7 +1,7 @@
 package com.daemawiki.external.web.rest.auth;
 
-import com.daemawiki.external.web.rest.auth.dto.LoginForm;
-import com.daemawiki.external.web.rest.auth.dto.RegisterForm;
+import com.daemawiki.external.web.rest.auth.dto.form.LoginForm;
+import com.daemawiki.external.web.rest.auth.dto.form.RegisterForm;
 import com.daemawiki.internal.core.domain.model.dto.auth.LoginDTO;
 import com.daemawiki.internal.core.domain.model.dto.auth.RegisterDTO;
 import org.mapstruct.Mapper;
