+++ src/main/java/com/daemawiki/external/web/rest/auth/LoginController.java
@@ -1,6 +1,6 @@
 package com.daemawiki.external.web.rest.auth;
 
-import com.daemawiki.external.web.rest.auth.dto.LoginForm;
+import com.daemawiki.external.web.rest.auth.dto.form.LoginForm;
 import com.daemawiki.internal.core.usecase.auth.LoginUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseCookie;
